move bottom app bar to components
