+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
 import './styles.css';
-import { BottomAppBar } from 'BottomAppBar';
+import { BottomAppBar } from 'components/BottomAppBar';
 
 export default function App() {
     return (
